removed autosummary signatures (#2359)
